refactor: 컴포넌트를 common과 setting-page 폴더로 구조화
