+++ src/pages/setting-page.tsx
@@ -3,10 +3,10 @@ import { useForm, FormProvider } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { formSchema, FormSchema } from "@/schemas/form-schema";
 import { FORM_FIELDS } from "@/constants/form-fields";
-import { FormSection } from "@/components/FormSection";
+import { FormSection } from "@/components/setting-page/FormSection";
 import { groupFieldsByCategory, sortFieldsByOrder } from "@/shared/utils/form-fields-utils";
-import { CustomButton } from "@/components/CustomButton";
-import { ParameterViewModal } from "@/components/ParameterViewModal";
+import { CustomButton } from "@/components/common/CustomButton";
+import { ParameterViewModal } from "@/components/setting-page/ParameterViewModal";
 
 const CATEGORY_SORT_CONFIG: Record<string, string[]> = {
   "Roll Pitch Yaw": ["Roll", "Pitch", "Yaw"],
